+++ sickrage/clients/qbittorrent.py
@@ -18,8 +18,9 @@
 
 from __future__ import unicode_literals
 
-import sickrage
 from requests.auth import HTTPDigestAuth
+
+import sickrage
 from sickrage.clients import GenericClient
 
 
@@ -39,7 +40,7 @@ class qbittorrentAPI(GenericClient):
         except Exception:
             return None
 
-        return self.auth if not self.response.status_code == 404 else None
+        return self.auth if self.response.ok else None
 
     def _add_torrent_uri(self, result):