+++ sickrage/providers/torrent/iptorrents.py
@@ -20,8 +20,9 @@ from __future__ import unicode_literals
 
 import re
 
-import sickrage
 from requests.utils import dict_from_cookiejar
+
+import sickrage
 from sickrage.core.caches.tv_cache import TVCache
 from sickrage.core.exceptions import AuthException
 from sickrage.core.helpers import bs4_parser, convert_size
@@ -70,7 +71,7 @@ class IPTorrentsProvider(TorrentProvider):
         login_params = {'username': self.username, 'password': self.password, 'login': 'submit'}
 
         response = sickrage.srCore.srWebSession.post(self.urls['login'], data=login_params, timeout=30)
-        if response.status_code != 200:
+        if not response.ok:
             sickrage.srCore.srLogger.warning("[{}]: Unable to connect to provider".format(self.name))
             return False